Customer Service, FoodDrink Service, Restaurant Service and Customer Service were updated.
I coded CRUD operations and etc for Customer, FoodDrink, Restaurant and Menu entities.

+++ onlineFoodOrdering/src/main/java/com/onlineFoodOrdering/onlineFoodOrdering/entity/Customer.java
@@ -34,7 +34,7 @@ public class Customer extends CommonUserKnowledge {
     }
 
     @JsonIgnore
-    private int totalSpendMoney;
+    private double totalSpendMoney;
 
     @JsonIgnore
     private int totalNumberOfOrder;
